requirements.txt
Django==1.6.1
argparse==1.2.1
django-ajax-selects==1.3.3
wsgiref==0.1.2
